// frontend/src/Enums/FiedTypes.ts
enum FieldTypes {
    TEXT = "text",
    EMAIL = "email",
    FILE = "file",
    SELECT = "select",
    SELECTCOLOR = "color",
    BOOLEAN = "boolean",
    REFERENCE = "reference",
    PASSWORD = "password",
    IMAGE = "image",
    TEXTAREA = "textarea",
    JSON = "json",
    RICHTEXT = "rich_text",
    DATE = "date",
    DATETIME = "datetime-local",
    SCRIPT = "script",
    TIMESTAMP = "timestamp",
    NUMBER = "number",
    FLOAT = "float",
    DOUBLE = "double",
    SELECTORADD = "select_or_add"
}

export default FieldTypes;