// frontend/src/Enums/TransactionCategory.ts
enum TransactionCategory {

    sell = "sell",
    devidend = "devidend",
    shareCapital = "share_capital",
    purchase = "purchase",
    otherIncom = "other_income",
    serviceFee = "service_fee",
    otherExpence = "other_expences",
    production = "production",
    tax = "tax",
    paidTax = "tax_paid"

};

export default TransactionCategory;